% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{multiline}
\alias{multiline}
\alias{multiline-package}
\title{multiline: an R package for reading multiline FWF files}
\description{
multiline: an R package for reading multiline FWF files
}
