% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiline.R
\name{read_multiline}
\alias{read_multiline}
\title{Read multiline FWF files}
\usage{
read_multiline(file, lines, col_positions, col_types = NULL, skip = 0, ...)
}
\value{
...
}
\description{
Read multiline FWF files
}
\examples{
NULL
}
\seealso{
...
}
